+++ examples/ecs.rs
@@ -12,12 +12,14 @@ fn main() -> Result<(), Box<dyn Error>> {
     );
 
     // call `DescribeRegions` with empty queries.
-    let response = aliyun_openapi_client.request("DescribeRegions", &[])?;
+    let response = aliyun_openapi_client.get("DescribeRegions").send()?;
     println!("DescribeRegions response: {}", response);
 
     // call `DescribeInstances` with queries.
-    let response =
-        aliyun_openapi_client.request("DescribeInstances", &[("RegionId", "cn-hangzhou")])?;
+    let response = aliyun_openapi_client
+        .get("DescribeInstances")
+        .query(&[("RegionId", "cn-hangzhou")])
+        .send()?;
     println!("DescribeInstances response: {}", response);
 
     Ok(())